Crate mpz_circuits

This crate provides types for representing computation as binary circuits.

Re-exports

pub use once_cell;

Modules

circuits

Pre-built circuits for MPC.

ops

Operations on binary encoded types.

types

Types for encoding other types as binary values.

Macros

evaluate

Evaluates a circuit and attempts to coerce the output into the specified return type indicated in the function signature.

test_circ

Helper macro for testing that a circuit evaluates to the expected value.

Structs

Circuit

A binary circuit.

CircuitBuilder

A circuit builder.

Tracer

A wrapper type for tracing operations applied to a value.

Enums

BuilderError

An error that can occur when building a circuit.

CircuitError

An error that can occur when performing operations with a circuit.

Gate

A binary logic gate.

GateType

The type of a gate.

Attribute Macros

trace

An attribute macro that can be applied to a function to automatically convert it into a circuit.