mpz_circuits
Attribute Macro trace

#[trace]

An attribute macro that can be applied to a function to automatically convert it into a circuit.

Example

use mpz_circuits::{trace, evaluate, CircuitBuilder};

#[trace]
fn bitxor(a: [u8; 16], b: [u8; 16]) -> [u8; 16] {
    std::array::from_fn(|i| a[i] ^ b[i])
}

fn main() {
    let builder = CircuitBuilder::new();
    let a = builder.add_array_input::<u8, 16>();
    let b = builder.add_array_input::<u8, 16>();

    let c = bitxor_trace(&mut builder.state(), a, b);

    builder.add_output(c);

    let circ = builder.build().unwrap();

    let a = [42u8; 16];
    let b = [69u8; 16];

    let output = evaluate!(circ, fn(a, b) -> [u8; 16]).unwrap();

    assert_eq!(output, bitxor(a, b));
}

Dependencies

Dependencies can be specified using the #[dep] attribute. This will replace any calls to the specified function with the provided trace function.

Path override

The default path of the trace is the original path appended with the _trace suffix.

This can be overridden by passing the path in as the second argument to the attribute, eg #[dep(old_path, new_path)].

Example

use mpz_circuits::{trace, evaluate, CircuitBuilder};

#[trace]
fn bitxor(a: [u8; 16], b: [u8; 16]) -> [u8; 16] {
   std::array::from_fn(|i| a[i] ^ b[i])
}

#[trace]
fn bitand(a: [u8; 16], b: [u8; 16]) -> [u8; 16] {
   std::array::from_fn(|i| a[i] & b[i])
}

#[trace]
#[dep(bitxor, bitxor_trace)]
#[dep(bitand)]
fn bitxor_and(a: [u8; 16], b: [u8; 16]) -> [u8; 16] {
   bitxor(a, bitand(a, b))
}

fn main() {
  let builder = CircuitBuilder::new();
  let a = builder.add_array_input::<u8, 16>();
  let b = builder.add_array_input::<u8, 16>();

  let c = bitxor_and_trace(&mut builder.state(), a, b);

  builder.add_output(c);

  let circ = builder.build().unwrap();

  let a = [42u8; 16];
  let b = [69u8; 16];

  let output = evaluate!(circ, fn(a, b) -> [u8; 16]).unwrap();

  assert_eq!(output, bitxor_and(a, b));
}

Cache

The macro can optionally be configured with the cache argument which will cache the circuit after the first invocation. This can be useful for functions that will be used multiple times.

The circuit will be cached for the lifetime of the program.

Suffix

The macro copies the traced function and appends the _trace suffix to the end of the name.

This preserves the original function, which can be used for testing.

This suffix can be overridden by passing the suffix = "new_suffix" argument to the macro.