mpz_circuits
Struct Tracer

pub struct Tracer<'a, T> { /* private fields */ }

A wrapper type for tracing operations applied to a value.

This type is used to track the operations applied to a value, which is used to build a circuit via a CircuitBuilder.

Implementations

impl<'a> Tracer<'a, U8>

pub fn from_be_bytes(bytes: [Tracer<'a, U8>; 1]) -> Tracer<'a, U8>

Create a value from its representation as a byte array in big endian.

pub fn to_be_bytes(self) -> [Tracer<'a, U8>; 1]

Returns the representation of this type as a byte array in big endian.

pub fn from_le_bytes(bytes: [Tracer<'a, U8>; 1]) -> Tracer<'a, U8>

Create a value from its representation as a byte array in little endian.

pub fn to_le_bytes(self) -> [Tracer<'a, U8>; 1]

Returns the representation of this type as a byte array in little endian.

impl<'a> Tracer<'a, U16>

pub fn from_be_bytes(bytes: [Tracer<'a, U8>; 2]) -> Tracer<'a, U16>

Create a value from its representation as a byte array in big endian.

pub fn to_be_bytes(self) -> [Tracer<'a, U8>; 2]

Returns the representation of this type as a byte array in big endian.

pub fn from_le_bytes(bytes: [Tracer<'a, U8>; 2]) -> Tracer<'a, U16>

Create a value from its representation as a byte array in little endian.

pub fn to_le_bytes(self) -> [Tracer<'a, U8>; 2]

Returns the representation of this type as a byte array in little endian.

impl<'a> Tracer<'a, U32>

pub fn from_be_bytes(bytes: [Tracer<'a, U8>; 4]) -> Tracer<'a, U32>

Create a value from its representation as a byte array in big endian.

pub fn to_be_bytes(self) -> [Tracer<'a, U8>; 4]

Returns the representation of this type as a byte array in big endian.

pub fn from_le_bytes(bytes: [Tracer<'a, U8>; 4]) -> Tracer<'a, U32>

Create a value from its representation as a byte array in little endian.

pub fn to_le_bytes(self) -> [Tracer<'a, U8>; 4]

Returns the representation of this type as a byte array in little endian.

impl<'a> Tracer<'a, U64>

pub fn from_be_bytes(bytes: [Tracer<'a, U8>; 8]) -> Tracer<'a, U64>

Create a value from its representation as a byte array in big endian.

pub fn to_be_bytes(self) -> [Tracer<'a, U8>; 8]

Returns the representation of this type as a byte array in big endian.

pub fn from_le_bytes(bytes: [Tracer<'a, U8>; 8]) -> Tracer<'a, U64>

Create a value from its representation as a byte array in little endian.

pub fn to_le_bytes(self) -> [Tracer<'a, U8>; 8]

Returns the representation of this type as a byte array in little endian.

impl<'a> Tracer<'a, U128>

pub fn from_be_bytes(bytes: [Tracer<'a, U8>; 16]) -> Tracer<'a, U128>

Create a value from its representation as a byte array in big endian.

pub fn to_be_bytes(self) -> [Tracer<'a, U8>; 16]

Returns the representation of this type as a byte array in big endian.

pub fn from_le_bytes(bytes: [Tracer<'a, U8>; 16]) -> Tracer<'a, U128>

Create a value from its representation as a byte array in little endian.

pub fn to_le_bytes(self) -> [Tracer<'a, U8>; 16]

Returns the representation of this type as a byte array in little endian.

impl<'a, T> Tracer<'a, T>

pub fn new(state: &'a RefCell<BuilderState>, value: T) -> Self

Create a new tracer.

pub fn to_inner(self) -> T

Return the inner value.

impl<'a> Tracer<'a, Bit>

pub fn node(&self) -> Node<Feed>

Returns the single node associated with the bit.