mpz_circuits
Module types

Types for encoding other types as binary values.

Structs

Bit

U8

U16

U32

U64

U128

Enums

BinaryRepr

A binary representation of a type.

TypeError

An error related to binary type conversions.

Value

A value that can be encoded into a binary representation.

ValueType

A value type that can be encoded into a binary representation.

Traits

BinaryLength

A type that has a constant bit length.

PrimitiveType

A primitive type.

StaticValueType

A type for which the value type can be statically determined.

ToBinaryRepr

A type that can be represented in binary form.